Add files via upload
A notebook to create a xyz file for tracking the evolution of structural units from large dump files (HISTORY). 
If you want to track the movement of the NbO6 octahedra in a LiNbO3, you want to pull the coordinates of the Nb-atoms you are interested at and the O-atoms bonded to them.